feat: change the repr() of DebianArchitectures
Typically the repr is used for non-user-facing things like logging and debugging;
however, there is a pervasive idiom in Craft codebases of using the string's repr
to pretty-print user-facing strings, where we want to show 'amd64' and not
'DebianArchitecture('amd64')'. Since a DebianArchiteture object *is* a string,
it's hard for client code to protect itself against "ugly reprs".

Since the two facts are in conflict, the decision is to update DebianArchiture's
repr to mimic the string's. The expectation is that this won't impact logging,
debugging and stack traces *too* much - if it does, we can revisit this change.